requirements.txt
google-genai==1.7.0
chromadb==0.6.3
google-api-core
pandas
streamlit
google-generativeai